Change convert.toint16 to convert.toint32

+++ Socket_XML_Send_Receive/Form1.cs
@@ -102,7 +102,7 @@ namespace Socket_XML_Send_Receive
             server1 = null;
             byte[] rcvBuffer_full = new byte[BUFSIZE_FULL];
             byte[] rcvBuffer_partial = new byte[BUFSIZE];
-            port_listen_int = System.Convert.ToInt16(textBox3.Text);
+            port_listen_int = System.Convert.ToInt32(textBox3.Text);
             using (server1 = new Socket(AddressFamily.InterNetwork, SocketType.Stream, ProtocolType.Tcp))
             {
                 try
